Add default values for test-prefix and test-suffix
Test file names are generated by appending project type specific
prefixes and suffixes to the implementation files. The functions
transforming project types to *fix strings return nil if a project
specific settings are not found. This commit changes the functions so
that a default string is always returned, which is better than raising
an error saying "project type not supported".

Adds two new defcustoms that define the default values for test-prefix
and test-suffix.